fix(hogql): normalize literal JSONExtract before lazy table resolution
Replace the bespoke JSONExtractString/JSONExtract -> materialized-column
rewrite in PropertySwapper with an early normalization pass that runs after
type resolution but before resolve_lazy_tables. It converts a literal
`JSONExtractString(properties, 'x')` into the same property-access Field that
`properties.x` produces, then lets the single existing resolution pipeline
(lazy tables -> lowering -> clickhouse_property_resolution) pick the physical
source.

Because the produced read now flows through lazy-table resolution like chain
access does, literal JSONExtract over the lazy persons/groups tables and over
JOINED-mode person.properties reach their materialized column inside the
generated subquery instead of decompressing the raw JSON blob. The bespoke
path could not (it ran after lazy resolution, when those tables were already
subqueries) and re-derived the physical column itself, which is what let the
table-name lookup drift from clickhouse_property_resolution.

Still gated on a static materialized column existing, so a call with no
materialized column keeps its raw-JSONExtract semantics unchanged.
