+++ posthog/hogql/transforms/property_types.py
@@ -6,7 +6,7 @@ from django.db.models.functions.comparison import Coalesce
 
 from posthog.hogql import ast
 from posthog.hogql.context import HogQLContext
-from posthog.hogql.database.models import BooleanDatabaseField, DateTimeDatabaseField, StringJSONDatabaseField
+from posthog.hogql.database.models import BooleanDatabaseField, DateTimeDatabaseField
 from posthog.hogql.database.s3_table import S3Table
 from posthog.hogql.database.schema.events import (
     EVENTS_TABLE_TYPES,
@@ -18,12 +18,10 @@ from posthog.hogql.database.schema.groups import GroupsTable
 from posthog.hogql.database.schema.persons import PersonsTable, RawPersonsTable
 from posthog.hogql.escape_sql import escape_hogql_identifier
 from posthog.hogql.property_planner import PropertySourceKind, plan_property_access
-from posthog.hogql.type_system import normalized_runtime_type, parse_sql_runtime_type
 from posthog.hogql.visitor import CloningVisitor, TraversingVisitor
 
 from posthog.clickhouse.materialized_columns import (
     DMAT_STRING_COLUMN_NAME_PREFIX,
-    MATERIALIZATION_VALID_TABLES,
     MaterializedColumn,
     TablesWithMaterializedColumns,
     get_materialized_column_for_property,
@@ -289,117 +287,11 @@ class PropertySwapper(CloningVisitor):
         )
 
     def visit_call(self, node: ast.Call):
-        rewritten = self._try_rewrite_json_extract_to_mat_column(node)
-        if rewritten is not None:
-            return rewritten
-
         self._inside_call_depth += 1
         try:
             return super().visit_call(node)
         finally:
             self._inside_call_depth -= 1
-
-    def _try_rewrite_json_extract_to_mat_column(self, node: ast.Call) -> ast.Field | None:
-        """Rewrite safe direct JSON property extraction to use a materialized column.
-
-        When users write raw JSONExtractString(properties, '$foo') in HogQL,
-        ClickHouse decompresses the full properties JSON blob. If '$foo' has a
-        materialized column (mat_$foo), this is unnecessary I/O. We rewrite the
-        call to a property access node that the printer resolves to the mat_ column.
-
-        Typed JSONExtract(...) calls are only rewritten when the physical column type
-        exactly matches the requested ClickHouse type, because JSON helper semantics
-        for missing keys and type mismatches differ by function family.
-        """
-        property_name = self._simple_json_extract_property_name(node)
-        if property_name is None:
-            return None
-
-        # Unwrap Alias if present (resolver wraps fields in Alias nodes)
-        field_arg = node.args[0]
-        if isinstance(field_arg, ast.Alias):
-            field_arg = field_arg.expr
-        if not isinstance(field_arg, ast.Field):
-            return None
-
-        # Unwrap FieldAliasType to get the underlying FieldType
-        field_type = field_arg.type
-        if isinstance(field_type, ast.FieldAliasType):
-            field_type = field_type.type
-        if not isinstance(field_type, ast.FieldType):
-            return None
-
-        database_field = field_type.resolve_database_field(self.context)
-        if not isinstance(database_field, StringJSONDatabaseField):
-            return None
-
-        table_type = field_type.table_type
-        while isinstance(table_type, (ast.TableAliasType, ast.ColumnAliasedTableType, ast.VirtualTableType)):
-            table_type = table_type.table_type
-        if not isinstance(table_type, ast.TableType):
-            return None
-
-        # The materialized-column registry is keyed by the ClickHouse table name, not the HogQL name
-        # (RawPersonsTable is "raw_persons" in HogQL but "person" in ClickHouse). Match clickhouse_property_resolution,
-        # or person/group properties read off `raw_persons`/`raw_groups` would miss their materialized column here.
-        table_name = table_type.resolve_database_table(self.context).to_printed_clickhouse(self.context)
-        if table_name not in MATERIALIZATION_VALID_TABLES:
-            return None
-
-        field_name = cast(TableColumn, database_field.name)
-        mat_col = get_materialized_column_for_property(
-            cast(TablesWithMaterializedColumns, table_name),
-            field_name,
-            property_name,
-        )
-        if mat_col is None:
-            return None
-
-        if not self._json_extract_matches_materialized_column_type(node, mat_col):
-            return None
-
-        return ast.Field(
-            start=node.start,
-            end=node.end,
-            chain=[*field_arg.chain, property_name],
-            type=ast.PropertyType(chain=[property_name], field_type=field_type),
-        )
-
-    @staticmethod
-    def _simple_json_extract_property_name(node: ast.Call) -> str | None:
-        if node.name == "JSONExtractString" and len(node.args) == 2:
-            prop_name_arg = node.args[1]
-        elif node.name == "JSONExtract" and len(node.args) == 3:
-            prop_name_arg = node.args[1]
-        else:
-            return None
-
-        if isinstance(prop_name_arg, ast.Constant) and isinstance(prop_name_arg.value, str):
-            return prop_name_arg.value
-        return None
-
-    @staticmethod
-    def _json_extract_matches_materialized_column_type(node: ast.Call, mat_col: MaterializedColumn) -> bool:
-        if node.name == "JSONExtractString":
-            # JSONExtractString has string semantics, so it only matches a string-backed column.
-            # A non-string materialized column (e.g. Nullable(Float64)) would otherwise be rewritten
-            # to the bare typed column, dropping the string type the surrounding query expects.
-            return parse_sql_runtime_type(mat_col.type).family == "string"
-
-        if node.name != "JSONExtract" or len(node.args) != 3:
-            return False
-
-        type_arg = node.args[2]
-        if not isinstance(type_arg, ast.Constant) or not isinstance(type_arg.value, str):
-            return False
-
-        # Normalize before comparing so formatting differences in the type spelling
-        # (whitespace, quoting) don't block the rewrite; semantic differences
-        # (nullability, width, timezone) still do, because JSON helper semantics for
-        # missing keys and out-of-range values differ from bare column semantics.
-        requested_type = normalized_runtime_type(parse_sql_runtime_type(type_arg.value))
-        materialized_type = normalized_runtime_type(parse_sql_runtime_type(mat_col.type))
-        return requested_type.family != "unknown" and requested_type == materialized_type
 
     def visit_compare_operation(self, node: ast.CompareOperation):
         result = super().visit_compare_operation(node)
