+++ posthog/hogql/transforms/test/test_property_types.py
@@ -495,16 +495,18 @@ class TestJSONExtractToMaterializedColumn(ClickhouseTestMixin, BaseTest):
             assert "pmat_$browser" in printed, f"Expected pmat_$browser in output, got: {printed}"
             assert "JSONExtractString" not in printed, f"Expected no JSONExtractString, got: {printed}"
 
-    def test_jsonextractstring_not_rewritten_on_lazy_persons_table(self):
-        # The lazy `persons` table expands into an argMax subquery, so its field_type is a LazyTableType, not a
-        # TableType, and the rewrite declines (a pmat_ reference here would be stranded outside the subquery). The
-        # materialized column is pulled in by the lazy-table machinery instead. Pins that boundary.
+    def test_jsonextractstring_rewritten_on_lazy_persons_table(self):
+        # Normalizing the call into a property-access read before lazy-table resolution lets it flow through the
+        # argMax subquery the same way `properties.$browser` chain access does, so the materialized column is read
+        # inside the subquery rather than the raw JSON blob. The two forms produce the same subquery (only the outer
+        # column alias differs), so the literal call now optimizes identically to chain access.
         with materialized("person", "$browser"):
-            printed = self._print_select("select JSONExtractString(properties, '$browser') from persons")
-            assert "JSONExtractString" in printed, (
-                f"Expected the lazy persons read to stay a JSON extract, got: {printed}"
-            )
-            assert "pmat_$browser" not in printed, f"Did not expect a bare pmat_ column, got: {printed}"
+            literal = self._print_select("select JSONExtractString(properties, '$browser') from persons")
+            chain = self._print_select("select properties.$browser from persons")
+            assert "pmat_$browser" in literal, f"Expected pmat_$browser in output, got: {literal}"
+            assert "JSONExtractString" not in literal, f"Expected no JSONExtractString, got: {literal}"
+            # Strip the outer SELECT column alias, which is all that differs (auto-derived vs `$browser`).
+            assert literal.split(" FROM ", 1)[1] == chain.split(" FROM ", 1)[1]
 
     def test_jsonextractstring_rewrites_all_calls_in_same_query(self):
         with materialized("events", "$browser"), materialized("events", "$os"):
